make offsets float

+++ api/models/base.py
@@ -30,11 +30,11 @@ class FrontendInput(BaseModel):
     dpi: float = Field(
         description="Indicates if SVG was created with Adobe Illustrator scaling or has a scaling in mm",
     )
-    x_offset: int = Field(
+    x_offset: float = Field(
         description="How many millimeters the svg should be moved alongside the x-axis",
         default=0,
     )
-    y_offset: int = Field(
+    y_offset: float = Field(
         description="How many millimeters the svg should be moved alongside the y-axis",
         default=0,
     )
